% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindingEligiblePatients.R
\name{EligiblePatients}
\alias{EligiblePatients}
\title{finding eligible patients in historical dataset}
\usage{
EligiblePatients(
  Registry,
  inclusion = NULL,
  exclusion = NULL,
  rm.na = F,
  summary = F
)
}
\arguments{
\item{Registry}{historical dataset}

\item{inclusion}{inclusion criteria, eg, "AGE >= 18"; multiple inclusion criteria, eg, list("AGE >= 18", "AGE < 64", "TEST1 + TEST2 > 50")}

\item{exclusion}{exclusion criteria, eg, "DISDUR > 36"; multiple inclusion criteria, eg, list("DISDUR > 36", "AGE >= 64")}

\item{rm.na}{remove na in the inclusion or exclusion variables, default as F}

\item{summary}{statistics of submitted full registry and eligible patients, default as F}
}
\value{
EligiblePatients
}
\description{
Data preparation for eligible historical dataset with inclusion and exclusion criteria
The function can select eligible patients with none, one or multiple inclusion or exclusion criteria. 
If neither inclusion nor exclusion criteria is provided, the function returns the original registry dataset.
}
\examples{
EligiblePatients (Registry = Registry, inclusion = list("AGE >= 18", "AGE < 64"), exclusion = "DISDUR > 36")
EligiblePatients (Registry = Registry, exclusion = "DISDUR > 36", rm.na = T)
EligiblePatients (Registry = Registry, summary = T)
}
