% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchingMI.R
\name{get.Matched}
\alias{get.Matched}
\title{Get Matches and test equivalence}
\usage{
get.Matched(
  formula,
  data.registry,
  data.rct,
  method = "nearest",
  ratio = 1,
  distance = "glm",
  id.var = "ALSNR",
  stime.var = "STIME_Death",
  status.var = "STATUS_Death",
  alpha = 0.2,
  plot = F,
  ...
)
}
\arguments{
\item{formula}{formula for matches}

\item{data.registry}{eligible censoring historical data for matching}

\item{data.rct}{Randomized Clinical Trial data}

\item{method}{method for propensity score matching; nearest by default}

\item{ratio}{match ratio; 1 by default}

\item{distance}{propensity score estimation method; glm by default; eg, lasso, randomforest ect.}

\item{id.var}{ID variable}

\item{stime.var}{survival time variable}

\item{status.var}{status variable}

\item{alpha}{alpha ratio, default as 0.2}

\item{plot}{F as no plot required; F by default; options: dist, caliper, spagetti, forest}
}
\value{
equivalence test rmst and hr

matched pairs and propensity score

matched dataset

plot-propensity score distribution, caliper distance, spagetti plot and forest plot
}
\description{
Get Matches based on formula and propensity score matching
}
\examples{
get.Matched (formula = as.formula (GRP ~ log (DISDUR) + as.factor(SEX) + BMI), data.registry, data.rct, method = "nearest", ratio = 1, distance = "glm", id.var = "ID", stime.var = "STIME_Death", status.var = "STATUS_Death", alpha = 0.2, plot = F)
get.Matched (formula = as.formula (GRP ~ log (DISDUR) + BMI), Registry = Registry, data.registry, data.rct, plot = "dist")
}
